# backend/docs/openapi.yaml
openapi: 3.0.2
servers:
  - url: http://localhost:8080/api/v1
    description: Production server
info:
  contact:
    email: 2002rostok@gmail.com
    name: R0st0k
  description: Information System "Post" API
  termsOfService: http://swagger.io/terms/
  title: Information System "Post"
  version: 0.7.0
tags:
  - name: Get Information
    description: Endpoints to get information
  - name: Update Information
    description: Endpoints to update information
paths:
  /sending:
    get:
      summary: Get information about a single sending
      description: >
        Get information about a sending by `order_id`.
        Require a complete match of `order_id`.
        Return `type`, `status` and `stages`.
      tags:
        - Get Information
      parameters:
        - $ref: '#/components/parameters/SendingOrderID'
      responses:
        '200':
          $ref: '#/components/responses/SendingGetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Registration of a new sending
      description: >
        Registration of a new sending.
        Require `type`, `sender`, `receiver`, `size`, `weight`.
        Return `order_id` of new sending.
      tags:
        - Update Information
      requestBody:
        $ref: '#/components/requestBodies/Sending'
      responses:
        '200':
          $ref: '#/components/responses/SendingPostResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /postcodes_by_settlement:
    get:
      summary: Get information about postcodes in cities
      description: >
        Get information about postcodes in cities.
        Return map with `settlement` key and `postcode` array value.
      tags:
        - Get Information
      parameters:
        - $ref: '#/components/parameters/PostOfficeTypeMultiple'
      responses:
        '200':
          $ref: '#/components/responses/PostcodesBySettlementGetResponse'
  /sending_filter:
    get:
      summary: Get sendings that fit the filter
      description: >
        Get sendings that fit the filter.
        Require `page` and `elems_on_page`.
        Return amount of sendings that fit the filter and sendings on the selected page.
      tags:
        - Get Information
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/ElementsOnPage'
        - $ref: '#/components/parameters/SendingFilterOrderID'
        - $ref: '#/components/parameters/SendingFilterType'
        - $ref: '#/components/parameters/SendingFilterStatus'
        - $ref: '#/components/parameters/SendingFilterDate'
        - $ref: '#/components/parameters/SendingFilterSettlements'
        - $ref: '#/components/parameters/SendingFilterLength'
        - $ref: '#/components/parameters/SendingFilterWidth'
        - $ref: '#/components/parameters/SendingFilterHeight'
        - $ref: '#/components/parameters/SendingFilterWeight'
        - $ref: '#/components/parameters/SendingSort'
      responses:
        '200':
          $ref: '#/components/responses/SendingFilterGetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /employee_filter:
    get:
      summary: Get employees that fit the filter
      description: >
        Get employees that fit the filter.
        Require `page` and `elems_on_page`.
        Return amount of employees that fit the filter and employees on the selected page.
      tags:
        - Get Information
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/ElementsOnPage'
        - $ref: '#/components/parameters/EmployeeFilterFullName'
        - $ref: '#/components/parameters/PostOfficeFilterSettlement'
        - $ref: '#/components/parameters/AddressPostcodeFilter'
        - $ref: '#/components/parameters/EmployeeFilterPosition'
        - $ref: '#/components/parameters/EmployeeFilterBirthDate'
        - $ref: '#/components/parameters/EmployeeFilterGender'
        - $ref: '#/components/parameters/EmployeeFilterPhoneNumber'
        - $ref: '#/components/parameters/EmployeeSort'
      responses:
        '200':
          $ref: '#/components/responses/EmployeeFilterGetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /data_export_sending:
    get:
      summary: Get data from collection `sendings` from database.
      description: >
        Get data from collection `sendings` from database.
        Return json array.
      tags:
        - Get Information
      responses:
        '200':
          $ref: '#/components/responses/DataExportSendingResponse'
  /data_import_sending:
    post:
      summary: Import data into collection `sendings` in database.
      description: >
        Import data into collection `sendings` in database.
        Require array of json.
      tags:
        - Update Information
      requestBody:
        $ref: '#/components/requestBodies/ImportSending'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  schemas:
    Error:
      type: object
      required:
        - error_message
      properties:
        error_message:
          type: string
          example: You don't look at errors anyway
    Sending:
      description: Schema of sending
      type: object
      required:
        - _id
        - order_id
        - registration_date
        - sender
        - receiver
        - type
        - size
        - weight
        - stages
        - status
      properties:
        _id:
          $ref: '#/components/schemas/ObjectID'
        order_id:
          $ref: '#/components/schemas/SendingOrderID'
        registration_date:
          description: Registration date of sending
          type: object
          required:
            - $date
          properties:
            $date:
              type: string
              format: date-time
        sender:
          $ref: '#/components/schemas/PostClient'
        receiver:
          $ref: '#/components/schemas/PostClient'
        type:
          $ref: '#/components/schemas/SendingType'
        size:
          $ref: '#/components/schemas/SendingSize'
        weight:
          $ref: '#/components/schemas/SendingWeight'
        stages:
          type: array
          items:
            $ref: '#/components/schemas/SendingStage'
        status:
          $ref: '#/components/schemas/SendingStatus'
    ObjectID:
      description: ObjectID of sending (24 byte hex string)
      type: object
      required:
        - $oid
      properties:
        $oid:
          type: string
          pattern: '^[a-fA-F0-9]{24}$'
    SendingOrderID:
      type: string
      format: uuid
      example: 123e4567-e89b-12d3-a456-426655440000
    SendingType:
      type: string
      enum: [ Письмо, Посылка, Бандероль ]
      example: Письмо
    SendingStatus:
      type: string
      enum: [ В пути, Доставлено, Утеряно ]
      example: В пути
    SendingSize:
      type: object
      required:
        - length
        - width
        - height
      properties:
        length:
          type: integer
          format: int64
          minimum: 10
          example: 100
        width:
          type: integer
          format: int64
          minimum: 10
          example: 73
        height:
          type: integer
          format: int64
          minimum: 1
          example: 42
    SendingWeight:
      type: integer
      format: int64
      minimum: 10
      example: 1000
    SendingStage:
      type: object
      required:
        - name
        - timestamp
        - postcode
        - employee_id
      properties:
        name:
          type: string
          enum:
            - Принято в отделении связи
            - Покинуло место приема
            - Прибыло в сортировочный центр
            - Покинуло сортировочный центр
            - Прибыло в место вручения
            - Вручено адресату
          example: Принято в отделении связи
        timestamp:
          type: object
          required:
            - $date
          properties:
            $date:
              type: string
              format: date-time
        postcode:
          $ref: '#/components/schemas/AddressPostcode'
        employee_id:
          $ref: '#/components/schemas/ObjectID'
    EmployeePosition:
      description: Available positions to employee
      type: string
      enum:
        - Сотрудник отделения связи
        - Сотрудник сортировочного центра
        - Водитель
        - Почтальон
      example: Сотрудник отделения связи
    EmployeeGender:
      description: Available gender of employee
      type: string
      enum: [ М, Ж ]
      example: М
    EmployeePhoneNumber:
      description: Phone number
      type: string
      pattern: '^\d{11}$'
      example: 89111234567
    Address:
      type: object
      required:
        - postcode
        - settlement
        - street
        - building
      properties:
        postcode:
          $ref: '#/components/schemas/AddressPostcode'
        region:
          type: string
          minLength: 1
          example: Республика Башкортостан
        district:
          type: string
          minLength: 1
          example: Мелеузовский район
        settlement:
          type: string
          minLength: 1
          example: пос. Нугуш
        street:
          type: string
          minLength: 1
          example: ул. Ленина
        building:
          type: string
          minLength: 1
          example: 42
        apartment:
          type: string
          minLength: 1
          example: 1
    AddressPostcode:
      type: string
      pattern: '^\d{6}$'
      example: 212017
    PostClient:
      type: object
      required:
        - name
        - surname
        - address
      properties:
        name:
          type: string
          minLength: 1
          example: Налсур
        surname:
          type: string
          minLength: 1
          example: Мулюков
        middle_name:
          type: string
          minLength: 1
          example: Рустэмович
        address:
          $ref: '#/components/schemas/Address'
    PostOfficeType:
      description: Available types of post offices
      type: string
      enum: [ Отделение связи, Сортировочный центр ]
      example: Отделение связи
    Page:
      type: integer
      format: int64
      minimum: 1
      example: 1
    ElementsOnPage:
      type: integer
      format: int64
      minimum: 1
      example: 5
    SendingSort:
      type: object
      properties:
        sort_type:
          type: string
          enum: [ asc, desc ]
          example: asc
        sort_field:
          type: string
          enum: [ order_id, type, status, date, sender_settlement, receiver_settlement, weight ]
          example: order_id
    EmployeeSort:
      type: object
      properties:
        sort_type:
          type: string
          enum: [ asc, desc ]
          example: asc
        sort_field:
          type: string
          enum: [ full_name, settlement, postcode, position, birth_date, gender, phone_number ]
          example: full_name
  parameters:
    SendingOrderID:
      in: query
      name: order_id
      description: ID of sending
      required: true
      schema:
        $ref: '#/components/schemas/SendingOrderID'
    Page:
      in: query
      name: page
      description: Current page
      required: true
      schema:
        $ref: '#/components/schemas/Page'
    ElementsOnPage:
      in: query
      name: elems_on_page
      description: The number of items displayed on the page
      required: true
      schema:
        $ref: '#/components/schemas/ElementsOnPage'
    SendingFilterOrderID:
      in: query
      name: order_id
      description: UUID or a part of it
      schema:
        type: string
        minLength: 1
        maxLength: 36
        example: 123e4567-e89b-12d3-a456-426655440000
    SendingFilterType:
      in: query
      name: type
      description: Array of filtered type of sending
      schema:
        type: array
        uniqueItems: true
        items:
          $ref: '#/components/schemas/SendingType'
        minItems: 1
        example: [ Письмо, Бандероль ]
    SendingFilterStatus:
      in: query
      name: status
      description: Array of filtered status of sending
      schema:
        type: array
        uniqueItems: true
        items:
          $ref: '#/components/schemas/SendingStatus'
        minItems: 1
        example: [ В пути, Доставлено ]
    SendingFilterDate:
      in: query
      name: date
      description: Sending registration date range
      schema:
        type: object
        properties:
          date_start:
            description: Start of sending date range
            type: string
            format: date
            example: 2022-11-20
          date_finish:
            description: Finish of sending date range
            type: string
            format: date
            example: 2022-11-23
    SendingFilterSettlements:
      in: query
      name: settlements
      description: Sender's or/and receiver's settlements or a part of them
      schema:
        type: object
        properties:
          sender_settlement:
            description: Sender's settlement or a part of it
            type: string
            minLength: 1
            example: Москва
          receiver_settlement:
            description: Receiver's settlement or a part of it
            type: string
            minLength: 1
            example: Санкт-Петербург
    SendingFilterLength:
      in: query
      name: length
      description: Sending length range or a part of range
      schema:
        type: object
        properties:
          length_min:
            type: integer
            format: int64
            minimum: 10
            example: 20
          length_max:
            type: integer
            format: int64
            minimum: 10
            example: 30
    SendingFilterWidth:
      in: query
      name: width
      description: Sending width range or a part of range
      schema:
        type: object
        properties:
          width_min:
            type: integer
            format: int64
            minimum: 10
            example: 20
          width_max:
            type: integer
            format: int64
            minimum: 10
            example: 30
    SendingFilterHeight:
      in: query
      name: height
      description: Sending height range or a part of range
      schema:
        type: object
        properties:
          height_min:
            type: integer
            format: int64
            minimum: 1
            example: 20
          height_max:
            type: integer
            format: int64
            minimum: 1
            example: 30
    SendingFilterWeight:
      in: query
      name: weight
      description: Sending weight range or a part of range
      schema:
        type: object
        properties:
          weight_min:
            $ref: '#/components/schemas/SendingWeight'
          weight_max:
            $ref: '#/components/schemas/SendingWeight'
    SendingSort:
      in: query
      name: sort
      description: Sending sort
      schema:
        $ref: '#/components/schemas/SendingSort'
    EmployeeFilterFullName:
      in: query
      name: full_name
      description: Full name of employee or a part of it
      schema:
        type: object
        properties:
          name:
            description: Name or a part of it
            type: string
            minLength: 1
            example: Ростислав
          surname:
            description: Surname or a part of it
            type: string
            minLength: 1
            example: Могилёвский
          middle_name:
            description: Middle name or a part of it
            type: string
            minLength: 1
            example: Сергеевич
    EmployeeFilterPosition:
      in: query
      name: position
      description: Array of filtered positions of employee
      schema:
        type: array
        uniqueItems: true
        minItems: 1
        items:
          $ref: '#/components/schemas/EmployeePosition'
    EmployeeFilterBirthDate:
      in: query
      name: birth_date
      description: Employee birth date range
      schema:
        type: object
        properties:
          birth_date_start:
            description: Start of employee birth date range
            type: string
            format: date
            example: 2022-11-20
          birth_date_finish:
            description: Finish of employee birth date range
            type: string
            format: date
            example: 2022-11-23
    EmployeeFilterGender:
      in: query
      name: gender
      description: Array of filtered gender of employee
      schema:
        type: array
        uniqueItems: true
        minItems: 1
        items:
          $ref: '#/components/schemas/EmployeeGender'
    EmployeeFilterPhoneNumber:
      in: query
      name: phone_number
      description: Phone number of employee or a part of it
      schema:
        type: string
        pattern: '^\d{1,11}$'
        example: 89111234567
    EmployeeSort:
      in: query
      name: sort
      description: Employee sort
      schema:
        $ref: '#/components/schemas/EmployeeSort'
    AddressPostcodeFilter:
      in: query
      name: postcode
      description: Full postcode
      schema:
        $ref: '#/components/schemas/AddressPostcode'
    PostOfficeFilterSettlement:
      in: query
      name: settlement
      description: Full name of post office's settlement or a part of it
      schema:
        type: string
        minLength: 1
        example: Санкт-Петербург
    PostOfficeTypeMultiple:
      in: query
      name: type
      description: Array of filtered type of post offices
      schema:
        type: array
        uniqueItems: true
        items:
          $ref: '#/components/schemas/PostOfficeType'
        minItems: 1
        example: [ Отделение связи, Сортировочный центр ]
  requestBodies:
    Sending:
      description: Information of sending
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - type
              - sender
              - receiver
              - size
              - weight
            properties:
              type:
                $ref: '#/components/schemas/SendingType'
              sender:
                $ref: '#/components/schemas/PostClient'
              receiver:
                $ref: '#/components/schemas/PostClient'
              size:
                $ref: '#/components/schemas/SendingSize'
              weight:
                $ref: '#/components/schemas/SendingWeight'
    ImportSending:
      description: Array of sending
      required: true
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Sending'
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    SendingGetResponse:
      description: ok
      content:
        application/json:
          schema:
            type: object
            required:
              - type
              - status
              - stages
            properties:
              type:
                $ref: '#/components/schemas/SendingType'
              status:
                $ref: '#/components/schemas/SendingStatus'
              stages:
                type: array
                items:
                  type: object
                  required:
                    - name
                    - date
                    - postcode
                    - settlement
                  properties:
                    name:
                      type: string
                      enum:
                        - Принято в отделении связи
                        - Покинуло место приема
                        - Прибыло в сортировочный центр
                        - Покинуло сортировочный центр
                        - Прибыло в место вручения
                        - Вручено адресату
                      example: Принято в отделении связи
                    date:
                      type: string
                      format: date-time
                    postcode:
                      $ref: '#/components/schemas/AddressPostcode'
                    settlement:
                      type: string
                      minLength: 1
                      example: Санкт-Петербург
    PostcodesBySettlementGetResponse:
      description: ok
      content:
        application/json:
          schema:
            type: object
            additionalProperties:
              type: array
              items:
                $ref: '#/components/schemas/AddressPostcode'
            example:
              Санкт-Петербург: [ 123001, 123002 ]
              Москва: [ 124001 ]
    SendingPostResponse:
      description: ok
      content:
        application/json:
          schema:
            type: object
            required:
              - order_id
            properties:
              order_id:
                $ref: '#/components/schemas/SendingOrderID'
    SendingFilterGetResponse:
      description: ok
      content:
        application/json:
          schema:
            type: object
            required:
              - total
              - result
            properties:
              total:
                type: integer
                format: int64
                minimum: 0
                example: 1
              result:
                type: array
                items:
                  type: object
                  required:
                    - order_id
                    - type
                    - date
                    - settlement
                    - weight
                    - size
                    - status
                  properties:
                    order_id:
                      $ref: '#/components/schemas/SendingOrderID'
                    type:
                      $ref: '#/components/schemas/SendingType'
                    date:
                      type: string
                      format: date
                      example: 2022-11-23
                    settlement:
                      type: object
                      required:
                        - sender
                        - receiver
                      properties:
                        sender:
                          type: string
                          example: Москва
                        receiver:
                          type: string
                          example: Санкт-петербург
                    weight:
                      $ref: '#/components/schemas/SendingWeight'
                    size:
                      $ref: '#/components/schemas/SendingSize'
                    status:
                      $ref: '#/components/schemas/SendingStatus'
    EmployeeFilterGetResponse:
      description: ok
      content:
        application/json:
          schema:
            type: object
            required:
              - total
              - result
            properties:
              total:
                type: integer
                format: int64
                minimum: 0
                example: 1
              result:
                type: array
                items:
                  type: object
                  required:
                    - name
                    - surname
                    - gender
                    - birth_date
                    - position
                    - phone_number
                    - settlement
                    - postcode
                  properties:
                    name:
                      type: string
                      minLength: 1
                      example: Антон
                    surname:
                      type: string
                      minLength: 1
                      example: Пустяков
                    middle_name:
                      type: string
                      minLength: 1
                      example: Олегович
                    gender:
                      $ref: '#/components/schemas/EmployeeGender'
                    birth_date:
                      type: string
                      format: date
                      example: 2022-11-23
                    position:
                      $ref: '#/components/schemas/EmployeePosition'
                    phone_number:
                      $ref: '#/components/schemas/EmployeePhoneNumber'
                    settlement:
                      type: string
                      minLength: 1
                      example: Санкт-Петербург
                    postcode:
                      $ref: '#/components/schemas/AddressPostcode'
    DataExportSendingResponse:
      description: ok
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Sending'